+++ Symfony/CS/Tests/Tokenizer/TokensTest.php
@@ -62,4 +62,49 @@ PHP;
 
         $this->assertSame($countBefore, $tokens->count());
     }
+
+    /**
+     * @dataProvider provideIsLambdaCases
+     */
+    public function testIsLambda($source, array $expected)
+    {
+        $tokens = Tokens::fromCode($source);
+
+        foreach ($expected as $index => $expected) {
+            $this->assertSame($expected, $tokens->isLambda($index));
+        }
+    }
+
+    public function provideIsLambdaCases()
+    {
+        return array(
+            array(
+                '<?php function foo () {}',
+                array(1 => false),
+            ),
+            array(
+                '<?php function /** foo */ foo () {}',
+                array(1 => false),
+            ),
+            array(
+                '<?php $foo = function () {}',
+                array(5 => true),
+            ),
+            array(
+                '<?php $foo = function /** foo */ () {}',
+                array(5 => true),
+            ),
+            array(
+                '<?php
+preg_replace_callback(
+    "/(^|[a-z])/",
+    function (array $matches) {
+        return "a";
+    },
+    $string
+);',
+                array(7 => true),
+            ),
+        );
+    }
 }